requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-anthropic>=0.1.0
anthropic>=0.25.0
google-generativeai>=0.5.0
PyGithub>=2.3.0
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
tenacity>=8.3.0
slowapi>=0.1.9
